remove unneded cast

+++ erizo/src/erizo/rtp/RtpPaddingGeneratorHandler.cpp
@@ -169,7 +169,7 @@ void RtpPaddingGeneratorHandler::recalculatePaddingRate(uint64_t target_padding_
     static_cast<int>(bytes_per_marker % (kMaxPaddingSize + rtp_header_length_) - rtp_header_length_);
   int clamped_payload_size =
     std::clamp(last_payload_size,
-    static_cast<int>(0), static_cast<int>(kMaxPaddingSize));
+      0, static_cast<int>(kMaxPaddingSize));
   last_padding_packet_size_ = static_cast<uint8_t>(clamped_payload_size);
 }
